Move ConfigTemplateLoader from utilities.jinja2 to extras.jinja2
